+++ CHANGELOG.rst
@@ -17,6 +17,12 @@ See `Semantic Versioning 2.0.0 <https://semver.org>`_.
 Releases and Important Milestones
 ---------------------------------
 
+PyPI 5.3.4 - TBD
+~~~~~~~~~~~~~~~~
+
+- updated README.rst
+  - previous version never fully updated for two circular array types
+
 PyPI 5.3.3 - 2025-09-21
 ~~~~~~~~~~~~~~~~~~~~~~~